requirements.txt
flask==2.0.1
flask-cors==3.0.10
youtube-transcript-api==0.4.4
deep-translator==1.9.1
murf
pydub==0.25.1
certifi==2023.7.22
Werkzeug==2.0.3
pyOpenSSL==23.2.0
cloudinary==1.37.0
python-dotenv==1.0.0
yt-dlp==2024.12.13
